+++ release-notes.md
@@ -3,6 +3,8 @@
 ## snapshot
 - ten receive addresses
 - wallet purpose, coin type and account
+- wallet history
+- electrum sockets
 
 ## 0.1.4
 - is mnemonic word valid